Combined execute policy and nav actions in a single script navigation.py
